Rename SpeciesFactory to GenomeFactory and added new params compatable to metadata genomeFactory module
